% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_initialism_dots}
\alias{add_initialism_dots}
\title{Add optional dots to initialisms}
\usage{
add_initialism_dots(x)
}
\description{
Convenience function for building regular expressions.
For example, \code{USA} can become \code{U.S.A}, \code{U.S.A.} and so on
}
