% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{clean_prescription_text}
\alias{clean_prescription_text}
\title{Clean up raw prescription freetext}
\usage{
clean_prescription_text(txt)
}
\arguments{
\item{txt}{a character vector}
}
\value{
a character vector the same length as \code{txt}
}
\description{
Clean up raw prescription freetext
}
\examples{
clean_prescription_text(example_prescriptions)

}
