% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_frequency.R
\name{convert_freq_text}
\alias{convert_freq_text}
\title{Convert daily frequency text into numeric form}
\usage{
convert_freq_text(x)
}
\arguments{
\item{string}{character vector output from \code{extract_number}

This is equivalent in some ways to \code{freq_word_con.py}.
The \code{frequency_dict} in the original Python code is implemented here via
the \code{word2num} utility function.}
}
\description{
Convert daily frequency text into numeric form
}
\examples{
convert_freq_text(c('daily', 'twice a day', '6 hrly', NA, 'weekly',
'1 - 2 a week', 'two to three wkly', 'every 2 weeks', '4 times weekly',
'3 - 4 times monthly', '5 - 6 times a month', '3 a week', '4 times a week',
 'two times at dinner', '1 before meals', 'one with each meal',
 'at mealtimes', 'take before meals to control blood sugar'))

library(tibble)
common_dosages <- read.csv('tests/common_doses_textmining.csv')
n <- 5000
x <- tibble(
  text = common_dosages[1:n, 'PRESCRIPTION'],
  old_min = common_dosages[1:n, 'DF.MIN'],
  old_max = common_dosages[1:n, 'DF.MAX'],
  dftxt = guess_frequency(text),
  dose_freq = convert_freq_text(dftxt)
)

}
