% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_number.R
\name{convert_number_text}
\alias{convert_number_text}
\title{Convert dose number text into numeric form}
\usage{
convert_number_text(x)
}
\arguments{
\item{string}{character vector output from \code{extract_number}

This is equivalent in some ways to \code{word_digit_con.py}, \code{word_digit_con2.py}}
}
\description{
Convert dose number text into numeric form
}
\examples{
convert_number_text(c('6 once per day', 'take five', 'nothing here', NA,
'2 three times at lunchtime and 3 at teatime'))

library(tibble)
common_dosages <- read.csv('tests/common_doses_textmining.csv')
x <- tibble(
  text = common_dosages[1:1000, 'PRESCRIPTION'],
  old_min = common_dosages[1:1000, 'DN.MIN'],
  old_max = common_dosages[1:1000, 'DN.MAX'],
  dntxt = guess_number(common_dosages[1:1000, 'PRESCRIPTION']),
  dose_num = convert_number_text(dntxt)
)

}
