% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cprd}
\alias{cprd}
\title{Sample electronic prescribing dataset}
\format{
An object of class \code{data.frame} with 714 rows and 6 columns.
}
\usage{
cprd
}
\description{
A dataset containing product codes, patient identifiers, quantities, dates
and free-text dose instructions, similar to data provided by the Clinical
Practice Research Datalink (CPRD).
}
\details{
Variables in the data include
\describe{
\item{id}{record identifier}
\item{patid}{patient identifier}
\item{date}{date of start of prescription}
\item{prodcode}{product code; identifier for the prescribed medication}
\item{qty}{total quantity of medication prescribed}
\item{text}{free text prescribing instructions}
}
}
\keyword{datasets}
