% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\name{dose_dict}
\alias{dose_dict}
\title{Dictionary of useful patterns for text mining prescriptions}
\usage{
dose_dict(x)
}
\arguments{
\item{x}{Name of category of keywords. If missing, returns whole dictionary

The possible categories of keywords include the following. If in doubt,
call \code{dose_dict()} with no argument to return the whole dictionary.
\itemize{
  \item unit
  \item numbers
  \item latin
}}
}
\description{
Dictionary of useful patterns for text mining prescriptions
}
\examples{
dose_dict('numbers')
}
