% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseminer-package.R
\docType{package}
\name{doseminer-package}
\alias{doseminer}
\alias{doseminer-package}
\title{doseminer: Dose Miner}
\description{
Extract dosage data from free-text prescriptions.
}
\author{
\strong{Maintainer}: David Antony Selby \email{david.selby@manchester.ac.uk} (\href{https://orcid.org/0000-0001-8026-5663}{ORCID})

Other contributors:
\itemize{
  \item Belay Birlie Yimer [contributor]
}

}
\keyword{internal}
