% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\docType{data}
\name{drug_units}
\alias{drug_units}
\title{Medication dosage units}
\format{
An object of class \code{character} of length 27.
}
\usage{
drug_units
}
\description{
A named character vector. Names represent patterns to match dose units
and values represent standardised names for those units.
}
\details{
Use with a function like \code{\link[stringr:str_replace]{str_replace_all}}
to standardise a freetext prescription.
}
\examples{
stringr::str_replace_all('Three milli litres', drug_units)
}
\keyword{datasets}
