% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cprd}
\alias{example_cprd}
\title{Example freetext prescriptions}
\format{
An object of class \code{character} of length 28.
}
\usage{
example_cprd
}
\description{
Adapted from CPRD common dosages
}
\seealso{
\code{\link{example_prescriptions}}
}
\keyword{datasets}
