% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{data}
\name{example_prescriptions}
\alias{example_prescriptions}
\title{Example freetext prescriptions}
\format{
An object of class \code{character} of length 18.
}
\usage{
example_prescriptions
}
\description{
Adapted from CPRD common dosages
}
\keyword{datasets}
