% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_number.R
\name{extract_dn_text}
\alias{extract_dn_text}
\title{Guess dosage number from prescription text}
\usage{
extract_dn_text(text)
}
\arguments{
\item{text}{Character vector of free text prescriptions}
}
\value{
Vector of substrings, to be passed to \code{guess_number}
}
\description{
Guess dosage number from prescription text
}
\examples{
presc <- c(
  '1-2 TABLET(S) UP TO FOUR TIMES A DAY WHEN REQUIRED',
  'THREE CAPSULES DAILY',
  'ONE EVERY DAY',
  'ONE TWICE A DAY',
  'TWICE A DAY',
  'TAKE ONE TWICE DAILY',
  '2 TABLETS FOUR TIMES A DAY',
  '5ML THREE TIMES DAILY WHEN REQUIRED',
  'TAKE 1 FOUR TIMES A DAY WHEN REQUIRED',
  '2 TABLETS FOUR TIMES A DAY WHEN REQUIRED',
  '1 CAPSULE TWICE A DAY',
  '5-10MLS FOUR TIMES A DAY WHEN REQUIRED',
  '2.5 OR 5ML FOUR TIMES A DAY WHEN REQUIRED',
  '10MG WHEN REQUIRED',
  '1 PATCH EVERY 72 HOURS'
)
stack(guess_dose_number(presc))

library(tibble)
common_dosages <- read.csv('tests/common_doses_textmining.csv')
x <- tibble(
  text = common_dosages[1:1000, 'PRESCRIPTION'],
  old_min = common_dosages[1:1000, 'DN.MIN'],
  old_max = common_dosages[1:1000, 'DN.MAX'],
  new = extract_dn_text(common_dosages[1:1000, 'PRESCRIPTION'])
)
}
