% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_dose_frequency_interval}
\alias{extract_dose_frequency_interval}
\title{Extract dose frequency information from freetext prescription}
\usage{
extract_dose_frequency_interval(txt)
}
\arguments{
\item{txt}{A character vector of freetext prescriptions}
}
\description{
Extract dose frequency information from freetext prescription
}
\examples{
extract_dose_frequency_interval(example_prescriptions)

}
