% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{extract_dose_unit}
\alias{extract_dose_unit}
\title{Extract units of dose from freetext prescriptions.}
\usage{
extract_dose_unit(txt)
}
\arguments{
\item{txt}{a character vector}
}
\value{
A character vector the same length as \code{txt}, containing
standardised units, or \code{NA} if no units were found in the prescription.
}
\description{
If there are multiple units in a string, only the first is returned.
}
\examples{
extract_dose_unit(example_prescriptions)

Based on \code{add_dose_unit.py} from original Python/Java algorithm.

}
