% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/new.R
\name{extract_drug_info}
\alias{extract_drug_info}
\title{Extract drug information}
\usage{
extract_drug_info(x)

extract_drug_info(x)
}
\description{
Based on \code{drug_information_extraction.py}.

Based on \code{drug_information_extraction.py}.
}
\details{
The aim of this function is to replicate the abovementioned Python script.
For now it will perform the following tasks:
\itemize{
\item Coerce to lower case
\item Trim leading and trailing whitespace
\item Replace pipes \code{|} with spaces (line 30)
\item Replace a sequence \code{numberxword} with \code{number x word}
\item Replace a sequence \code{numberword} with \code{number word}
\item Replace a sequence \code{wordnumberword} with \code{word number word}
\item Replace a sequence \code{numberwordnumber} with \code{number word number}
\item Split up sequences with units (see line 40 of \code{.py} script)
}
We have also added some stuff that may not have been in the original script:
replacing double spaces, tabs or newline characters with single spaces.

The aim of this function is to replicate the abovementioned Python script.
For now it will perform the following tasks:
\itemize{
\item Coerce to lower case
\item Trim leading and trailing whitespace
\item Replace pipes \code{|} with spaces (line 30)
\item Replace a sequence \code{numberxword} with \code{number x word}
\item Replace a sequence \code{numberword} with \code{number word}
\item Replace a sequence \code{wordnumberword} with \code{word number word}
\item Replace a sequence \code{numberwordnumber} with \code{number word number}
\item Split up sequences with units (see line 40 of \code{.py} script)
}
We have also added some stuff that may not have been in the original script:
replacing double spaces, tabs or newline characters with single spaces.
}
\examples{
extract_drug_info(c('2mg per day', '1xdaily', 'a2b', '  double  spaced  ',
                    'newline\ntab\treturn\r', '2by14', 'take q4d'))

extract_drug_info(c('2mg per day', '1xdaily', 'a2b', '  double  spaced  ',
                    'newline\ntab\treturn\r', '2by14', 'take q4d'))

}
