% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_from_prescription}
\alias{extract_from_prescription}
\title{Extract dose frequency information from freetext prescription}
\usage{
extract_from_prescription(txt)
}
\arguments{
\item{txt}{A character vector of freetext prescriptions}
}
\description{
This is the main workhorse function for the \code{doseminer} package.
}
\examples{
extract_from_prescription(example_prescriptions)

}
