% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_number.R
\name{extract_number_text}
\alias{extract_number_text}
\title{Extract text corresponding to dosage number}
\usage{
extract_number_text(text)
}
\arguments{
\item{text}{Character vector of free text prescriptions}
}
\value{
Vector of substrings, to be passed to \code{guess_number}
}
\description{
Seeing if we can do better than original without copying over code.
}
\examples{
library(tibble)
common_dosages <- read.csv('tests/common_doses_textmining.csv')
x <- tibble(
  text = common_dosages[1:1000, 'PRESCRIPTION'],
  old_min = common_dosages[1:1000, 'DN.MIN'],
  old_max = common_dosages[1:1000, 'DN.MAX'],
  new = extract_number_text(common_dosages[1:1000, 'PRESCRIPTION'])
)
}
