% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_number.R
\name{guess_dose_number}
\alias{guess_dose_number}
\title{Guess dosage number from prescription text}
\usage{
guess_dose_number(text)
}
\arguments{
\item{text}{character vector of raw prescription text}
}
\description{
Guess dosage number from prescription text
}
\examples{
presc <- c(
  '1-2 TABLET(S) UP TO FOUR TIMES A DAY WHEN REQUIRED',
  'THREE CAPSULES DAILY',
  'ONE EVERY DAY',
  'ONE TWICE A DAY',
  'TWICE A DAY',
  'TAKE ONE TWICE DAILY',
  '2 TABLETS FOUR TIMES A DAY',
  '5ML THREE TIMES DAILY WHEN REQUIRED',
  'TAKE 1 FOUR TIMES A DAY WHEN REQUIRED',
  '2 TABLETS FOUR TIMES A DAY WHEN REQUIRED',
  '1 CAPSULE TWICE A DAY',
  '5-10MLS FOUR TIMES A DAY WHEN REQUIRED',
  '2.5 OR 5ML FOUR TIMES A DAY WHEN REQUIRED',
  '10MG WHEN REQUIRED',
  '1 PATCH EVERY 72 HOURS'
)
stack(guess_dose_number(presc))

}
