% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{guess_frequency}
\alias{guess_frequency}
\title{Guess frequency of dose}
\usage{
guess_frequency(text)
}
\description{
Guess frequency of dose
}
\examples{
guess_frequency(
c('5 daily', 'four monthly', '5 q4h', 'two x3hly', 'take 1 weekly', # NB initial verb
  '3 / day', '2/week', '3x per day', 'five q2h', 'three am', 'one daily'))

x <- data.frame(text = common_dosages[1:1000, 'PRESCRIPTION'],
           old_min = common_dosages[1:1000, 'DF.MIN'],
           old_max = common_dosages[1:1000, 'DF.MAX'],
           new = unname(guess_frequency(common_dosages[1:1000, 'PRESCRIPTION'])))
}
