% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{guess_number}
\alias{guess_number}
\title{Guess dosage number (rewriting from scratch)}
\usage{
guess_number(text)
}
\description{
Seeing if we can do better than original without copying over code.
}
\examples{
n <- 5000
x <- data.frame(text = common_dosages[1:n, 'PRESCRIPTION'],
           old_min = common_dosages[1:n, 'DN.MIN'],
           old_max = common_dosages[1:n, 'DN.MAX'],
           new = unname(guess_number(common_dosages[1:n, 'PRESCRIPTION'])))
}
