% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{guess_prescription}
\alias{guess_prescription}
\title{Guess dosage from prescription free text}
\usage{
guess_prescription(text)
}
\description{
Guess dosage from prescription free text
}
\examples{
View(guess_prescription(common_dosages[1:1000, 'PRESCRIPTION']))
}
