% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{hourly_to_daily}
\alias{hourly_to_daily}
\title{Convert hourly to daily frequency}
\usage{
hourly_to_daily(txt)
}
\arguments{
\item{txt}{String of the form 'every n hours'}
}
\description{
Convert hourly to daily frequency
}
\examples{
\dontrun{
hourly_to_daily('every 1 hour')
hourly_to_daily('every 4 hours')
hourly_to_daily('every 3 - 4 hours')
hourly_to_daily('every 36 - 72 hours')
}

}
