% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{multiply_dose}
\alias{multiply_dose}
\title{Evaluate a multiplicative plaintext expression}
\usage{
multiply_dose(axb)
}
\description{
Replaces written phrases like "2 x 5" with their arithmetic result (i.e. 10)
}
\examples{
multiply_dose('2 x 5')
multiply_dose('2 - 4 x 5')

}
\seealso{
Used internally within \code{\link{extract_from_prescription}}
}
