% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{parse_frequency}
\alias{parse_frequency}
\title{Convert extracted frequency text into numeric daily frequency of dose}
\usage{
parse_frequency(string)
}
\arguments{
\item{string}{Text extracted by \code{\link{guess_frequency}} (probably should rename this \code{extract_} or \code{match_})}
}
\value{
Estimated (minimum and maximum) times per day dose is administered, as numeric value.
}
\description{
Based on the script \code{freq_word_con.py} in original algorithm.
}
