% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\docType{data}
\name{patterns}
\alias{patterns}
\title{Regular expression patterns for prescription text mining}
\format{
An object of class \code{list} of length 12.
}
\usage{
patterns
}
\description{
Mainly for internal use. You can list all available patterns with
\code{names(doseminer:::patterns)}
}
\keyword{datasets}
