% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regex_or}
\alias{regex_or}
\alias{regex_data_or}
\title{Build regular expressions from vectors of alternatives}
\usage{
regex_or(..., .envir = parent.frame())

regex_data_or(.x, ..., .envir = parent.frame())
}
\arguments{
\item{...}{Arguments passed to \code{\link[glue]{glue}}}
}
\description{
A light wrapper around \code{\link[glue]{glue}}.
}
\details{
Any expressions of the form \code{\{foo*\}} will be interpreted like
\code{\{paste(foo, collapse = '|')\}} when the string is interpolated.
}
\examples{
ntimes <- c('once', 'twice', 'thrice')
regex_or('Take {ntimes*} per day')

words <- list(n = ntimes)
regex_data_or(words, 'Take ({n*}) daily')

# Easier to read but maybe less efficient:
library(glue)
glue_collapse(glue('Take {ntimes} per day'), sep = '|')

}
