% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{weekly_to_daily}
\alias{weekly_to_daily}
\title{Convert weekly interval to daily interval}
\usage{
weekly_to_daily(Dperweek)
}
\arguments{
\item{Dperweek}{String of the form 'n / week'}
}
\value{
An equivalent string of the form 'x / day'
}
\description{
Convert weekly interval to daily interval
}
