% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{word2num}
\alias{word2num}
\title{Coerce English number words to digits}
\usage{
word2num(x)
}
\arguments{
\item{x}{a character vector of words}
}
\value{
a vector the same length as \code{x}, with any named numbers
converted to the respective digits.
}
\description{
If a number is given in word form, for example 'three', it is converted to
digits, i.e. '3' corresponding to a daily dosage.
}
\details{
Current implementation covers numbers one to twenty four, plus common Latin
abbreviations (e.g. q.q.h. means 'every four hours', i.e. 6 per day).
}
\examples{
word2num(c('three', 'six', 'hello', 'q4h', 'bedtime'))

}
