% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{words2number}
\alias{words2number}
\title{Convert English names of numbers to their numerical values}
\source{
Originally adapted from the
\href{https://github.com/benmarwick/words2number}{\code{words2number}} package by
Ben Marwick.
}
\usage{
words2number(txt)
}
\arguments{
\item{txt}{A character vector containing names of numbers (only).}
}
\value{
A named numeric vector of the same length as \code{phrase}.
}
\description{
Convert English names of numbers to their numerical values
}
\examples{
words2number('seven')
words2number('forty-two')
words2number(c('three', 'one', 'twenty two thousand'))

}
