% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseminer-package.R
\docType{package}
\name{doseminer-package}
\alias{doseminer}
\alias{doseminer-package}
\title{doseminer: Extract Drug Dosages from Free-Text Prescriptions}
\description{
Utilities for converting unstructured electronic prescribing instructions into structured medication data. Extracts drug dose, units, daily dosing frequency and intervals from English-language prescriptions. Based on Karystianis et al. (2015) <doi:10.1186/s12911-016-0255-x>.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/Selbosh/doseminer/issues}
}

}
\author{
\strong{Maintainer}: David Selby \email{david.selby@manchester.ac.uk} (\href{https://orcid.org/0000-0001-8026-5663}{ORCID})

Other contributors:
\itemize{
  \item Belay Birlie Yimer \email{belaybirlie.yimer@manchester.ac.uk} [contributor]
  \item Ben Marwick \email{benmarwick@gmail.com} [contributor]
}

}
\keyword{internal}
