% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_prescriptions}
\alias{example_prescriptions}
\title{Example freetext prescriptions}
\format{
An object of class \code{character} of length 27.
}
\usage{
example_prescriptions
}
\description{
Various examples of how prescription data may be represented in free text.
}
\seealso{
\code{\link{example_cprd}}
}
\keyword{datasets}
