% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{guess_frequency}
\alias{guess_frequency}
\title{Guess frequency of dose}
\usage{
guess_frequency(text)
}
\description{
The frequency of dose means how many times, on 'dose day', that a dose
should be administered. Even if there are several days between doses, and
even if each actual dose contains multiple units. Thus 'two in the morning'
is a frequency of 1 (because you take both at once). Similarly '1 weekly'
is also a frequency of 1, because though there are gaps between dose days,
on the day of the actual dose, one is administered.
}
\examples{
guess_frequency(
c('5 daily', 'four monthly', '5 q4h', 'two x3hly', 'take 1 weekly', # NB initial verb
  '3 / day', '2/week', '3x per day', 'five q2h', 'three am', 'one daily'))

library(tibble)
n <- 5000
x <- tibble(text = common_dosages[1:n, 'PRESCRIPTION'],
           old_min = common_dosages[1:n, 'DF.MIN'],
           old_max = common_dosages[1:n, 'DF.MAX'],
           txt_freq = unname(guess_frequency(text)),
           num_freq = parse_frequency(txt_freq)
)
View(x)
}
