% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{hourly_to_daily}
\alias{hourly_to_daily}
\title{Convert hourly to daily frequency}
\usage{
hourly_to_daily(txt)
}
\arguments{
\item{txt}{String of the form 'every n hours'}
}
\value{
An equivalent string of the form 'x / day'
}
\description{
Convert hourly to daily frequency
}
