% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{hourly_to_daily}
\alias{hourly_to_daily}
\title{Convert hourly to daily frequency}
\usage{
hourly_to_daily(everyDhrs)
}
\description{
Convert hourly to daily frequency
}
\examples{
hourly_to_daily('every 4 hours')

}
