% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{multiply_dose}
\alias{multiply_dose}
\title{Evaluate a multiplicative plaintext expression}
\usage{
multiply_dose(axb)
}
\arguments{
\item{axb}{An string expression of the form 'A x B' where A, B are numeric}
}
\value{
An equivalent string giving the product of \code{A} and \code{B}.
If \code{A} is a range of values, a range of values is returned.
}
\description{
Replaces written phrases like "2 x 5" with their arithmetic result (i.e. 10)
}
\seealso{
Used internally within \code{\link{extract_from_prescription}}
}
